+++ arcane/src/arcane/accelerator/core/Runner.cc
@@ -19,13 +19,13 @@
 #include "arcane/utils/ArgumentException.h"
 #include "arcane/utils/MemoryView.h"
 
-#include "arcane/accelerator/core/RunQueueImpl.h"
 #include "arcane/accelerator/core/RunQueueBuildInfo.h"
 #include "arcane/accelerator/core/DeviceId.h"
 #include "arcane/accelerator/core/IDeviceInfoList.h"
 #include "arcane/accelerator/core/PointerAttribute.h"
 #include "arcane/accelerator/core/internal/IRunnerRuntime.h"
 #include "arcane/accelerator/core/internal/AcceleratorCoreGlobalInternal.h"
+#include "arcane/accelerator/core/internal/RunQueueImpl.h"
 
 #include <stack>
 #include <map>
